Move Colors component to docs dir and remove redundnat PartialTokens type.
